[compiler][newinference] ensure fixpoint converges for loops w backedges
The fixpoint converges based on the abstract values being equal, but we synthesize new values cached based on the effects and new effects can be synthesized on the fly. So here we intern the effect objects by "value" (cache key computed from the value) to ensure that effects are stable, values cached based on them are stable, and that the fixpoint can converge.

ghstack-source-id: 9ca53d7e834407a0d4ca1edc850d9355984ccca7
Pull Request resolved: https://github.com/facebook/react/pull/33449