Update base for Update on "[compiler][newinference] ensure fixpoint converges for loops w backedges"
The fixpoint converges based on the abstract values being equal, but we synthesize new values cached based on the effects and new effects can be synthesized on the fly. So here we intern the effect objects by "value" (cache key computed from the value) to ensure that effects are stable, values cached based on them are stable, and that the fixpoint can converge.

[ghstack-poisoned]