[compiler] New inference repros/fixes
Adds a mechanism where if we can't determine a precise set of aliasing (data flow) effects for a function signature, to fall back to treating function expressions as general-purpose mutable values. Concretely:

* `InferMutationAliasingEffects` has an optimization where if you're `Apply`-ing a specific, known function expression, that we use the inferred signature for that function. Now we only use this mode _if_ we could infer a precise signature.
* `InferFunctionExpressionAliasingSignature` now bails out if there is a phi with a backedge, since we weren't handling this case.

See the repro for this case: i confirmed that it was previously incorrect.

Also adds a repro for a case of an invalid array length, which exposes the fact that InferFunctionExpressionAliasingSignatures is accumulating aliases w/o reducing them in any way. A complex enough function expression can lead to the accumulated array of effects overflowing the max array size. Fix in the next PR.