[compiler] Add new ValidateNoVoidUseMemo pass
Adds a new validation pass to validate against `useMemo`s that don't return anything. This usually indicates some kind of "useEffect"-like code that has side effects that need to be memoized to prevent overfiring, and is an anti-pattern.

A follow up validation could also look at the return value of `useMemo`s to see if they are being used.